==> N_uptake_NEON/01_download_NEON_data.qmd <==
---
title: "00_download_NEON_data"
format: html
editor: visual
---

## Download NEON data

part of the space-time-rivers NO3 sensor data project. download and visualize R data from NEON streams.

Code to DL: Nitrate in surface water: DP1.20033.001

Other useful data: Continuous discharge: DP4.00130.001 Discharge field collection: DP1.20048.001 (manual Q data) PAR: DP1.00024.001 PAR at water surface: DP1.20042.001 PAR below water surface: DP1.20261.001 Precipitation DP1.00006.001 Temp. in surface water: DP1.20053.001 Water quality DP1.20288.001 (SC, Chl-a, DO content, fDOM concentration, pH, and turbidity) \#

#### Load packages

```{r - load packages, NEON token}

# load packages
library(neonUtilities)
library(neonOS)
library(raster)
library(tidyverse)
library(here)

# set NEON token (FLBS email one)
NEON_TOKEN <- 'eyJ0eXAiOiJKV1QiLCJhbGciOiJFUzI1NiJ9.eyJhdWQiOiJodHRwczovL2RhdGEubmVvbnNjaWVuY2Uub3JnL2FwaS92MC8iLCJzdWIiOiJjaHJpc3RhLnRvcnJlbnNAZmxicy51bXQuZWR1Iiwic2NvcGUiOiJyYXRlOnB1YmxpYyIsImlzcyI6Imh0dHBzOi8vZGF0YS5uZW9uc2NpZW5jZS5vcmcvIiwiZXhwIjoxODMyMzUwMzM5LCJpYXQiOjE2NzQ2NzAzMzksImVtYWlsIjoiY2hyaXN0YS50b3JyZW5zQGZsYnMudW10LmVkdSJ9.9LnYeA5NuTuCSeZkhQwizdh07i9dK4zmUYraXQwhrfkvaF75KNFJnqp04qJNZyhYQZFO3rHhUx7FJIsDY5sVcw'


# Set global option to NOT convert all character variables to factors
options(stringsAsFactors=F)

# If you are working in R version 4 or higher (recommended), this is already the default setting.


```

#### Download data files and save as Rdata

Bobby Hensley informed us that Q and O2 data are unreliable before 2021 (WY2022!). Currently (1/16/24) NEON NO3 data are provisional from July 2023 on. Downloading all of 2021-2023, and will run again once data are in final form.

```{r - DL and save data files}

### download NO3 and related datasets  ###
# sitelist <- c("ARIK", "BIGC", "BLDE", "BLUE", "BLWA", "CARI", "COMO", "CUPE", "FLNT", "GUIL","HOPB", "KING", "LECO", "MART", "OKSR", "POSE", "PRIN", "REDB", "SYCA", "TECR", "WALK", "WLOU")  #already have "CARI", "BIGC", "KING", "WALK", but re-downloading so everything is together

sitelist <- c("BIGC", "BLDE", "CARI", "CUPE", "MART", "OKSR", "PRIN", "REDB", "SYCA", "WLOU")

# sitelist <- "OKSR"

# Not pulled:Lewis Run [LEWI], Clarke, VA; Mayfield Creek [MAYF], Bibb, AL; McDiffett Creek [MCDI], Wabaunsee, KS; McRae Creek [MCRA], Linn, OR; Tombigbee River [TOMB], Choctaw, AL

# startdate and enddate need to be character strings in YYYY-MM format 
# BUT, these are and it didn't run/ threw errors that the format is incorrect!

# startdate <- 2021-01
# enddate <- 2023-12

no3.data.2124 <- loadByProduct(dpID = "DP1.20033.001",
                          site = sitelist,
                          startdate="2021-10", enddate="2024-09",  # 2024 NO3 has been QA/QC'd
                          include.provisional=TRUE,
                          package="expanded", check.size=F,
                          token = Sys.getenv("NEON_TOKEN"))

# # PAR is HUGE. Wait on this/ do individually? Store in cloud?
# par_wsurf_data <- loadByProduct(dpID = "DP1.20042.001",
#                                 # site = sitelist,
#                                 site = c("CARI", "OKSR", "BIGC", "WALK"),
#                                 startdate="2021-01", enddate="2024-06",
#                                 include.provisional=FALSE,
#                                 package="expanded", check.size=F,
#                                 token = Sys.getenv("NEON_TOKEN"))


par_data <- loadByProduct(dpID = "DP1.00024.001",
                                # site = sitelist,
                                startdate="2021-10", enddate="2024-09",
                                site = sitelist,
                                include.provisional=FALSE,
                                package="expanded", check.size=F,
                                token = Sys.getenv("NEON_TOKEN"))


# precip_data <- loadByProduct(dpID = "DP1.00006.001", 
#                             site = sitelist, 
#                             package="expanded", check.size=F, 
#                             token = Sys.getenv("NEON_TOKEN"))
# 
q_data <- loadByProduct(dpID = "DP4.00130.001",
                        site = sitelist,
                        startdate="2021-10", enddate="2024-09",  
                        # per Bobby H. the WY24 Q data are cleaned and ready to use - 8/7/2025
                        include.provisional=TRUE,
                        package="expanded", check.size=F,
                        token = Sys.getenv("NEON_TOKEN"))

# q_data.prov <- loadByProduct(dpID = "DP4.00130.001",
#                         site = sitelist,
#                         startdate="2023-10", enddate="2024-12",
#                         include.provisional=TRUE,
#                         package="expanded", check.size=F,
#                         token = Sys.getenv("NEON_TOKEN"))
# 
# manual.q_data <- loadByProduct(dpID = "DP1.20048.001",              # field discharge data
#                         site = sitelist, 
#                         package="expanded", check.size=F, 
#                         token = Sys.getenv("NEON_TOKEN"))
# 
# wtemp_data <- loadByProduct(dpID = "DP1.20053.001", 
#                             site = sitelist, 
#                             package="expanded", check.size=F, 
#                             token = Sys.getenv("NEON_TOKEN"))
# 
#  
wqual_data <- loadByProduct(dpID = "DP1.20288.001",
                            site = sitelist,
                            startdate="2023-10", enddate="2024-09",
                            include.provisional=FALSE,
                            package="expanded", check.size=F,
                            token = Sys.getenv("NEON_TOKEN"))


# reaeration_data <- loadByProduct(dpID = "DP1.20190.001",  # reaeration field and lab data
#                         site = sitelist, 
#                         package="expanded", check.size=F, 
#                         token = Sys.getenv("NEON_TOKEN"))

### Save to data_raw ###

save(no3.data.2124, file=here("N_uptake_NEON/data/neon_data_raw/no3_data_all_2124.Rdata"))

save(par_wsurf_data, file=here("N_uptake_NEON/data/neon_data_raw/par_wsurf_data.Rdata"))

save(par_data, file=here("N_uptake_NEON/data/neon_data_raw/par_data_OKSR.Rdata"))

# save(precip_data, file=here("N_uptake_NEON/data/neon_data_raw/precip_data.Rdata"))
save(q_data, file= here("N_uptake_NEON/data/neon_data_raw/q_data_2124.Rdata"))
save(q_data.prov, file= here("N_uptake_NEON/data/neon_data_raw/q_data_2324prov.Rdata"))
# save(manual.q_data, file=here("N_uptake_NEON/data/neon_data_raw/manualq_data.Rdata"))
# save(wtemp_data, file=here("N_uptake_NEON/data/neon_data_raw/wtemp_data.Rdata"))
# save(wqual_data, file=here("N_uptake_NEON/data/neon_data_raw/wqual_data.Rdata"))
# save(reaeration_data, file=here("N_uptake_NEON/data/neon_data_raw/reaeration_data.Rdata"))


```

### Loop to download large WQual files

Run packages + NEON_TOKEN 1st

```{r loop for lg files}

# Downloads water quality data by year and site; saves to GPP_model/ WQ_data_NEON_raw folder

sitelist <- c("BIGC", "BLDE", "CARI", "CUPE", "MART", "OKSR", "PRIN", "REDB", "SYCA")

years <- 2021:2024

for (site in sitelist) {
  for (year in years) {
    message(paste("Downloading", site, year))
    
    temp <- try(
      loadByProduct(
        dpID = "DP1.20288.001",
        site = site,
        startdate = paste0(year, "-01"),
        enddate = paste0(year, "-12"),
        package = "expanded",
        token = Sys.getenv("NEON_TOKEN"),
        check.size = FALSE
      ), silent = TRUE
    )

      # Save filtered data to your path
      saveRDS(
        temp,
        file = here("N_uptake_NEON", "GPP_model", "DO_data_neon_raw",
                    paste0("WQual_", site, "_", year, ".rds"))
      )
    
    rm(temp); gc()
  }
}




```
